code/requirements.txt
cassandra-driver==3.24.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
geomet==0.2.1.post1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
pandas==1.2.3
pika==1.2.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
watchdog==2.0.2
Werkzeug==1.0.1
